Add Config for dry run mode
A dry run mode was suggested to go in early[1], and I think that's an
extremely sensible idea!

To begin with, I've added a simple flag that will be loaded when the
framework is imported using `pflag`[2], so the user just needs to
specify `--dry-run`.

This actually creates a whole new concept of `Config`, which could be
anything that we want to specify when we run code built with the
framework.

For the time-being I've added a simple quick return to each function.

[1] https://github.com/surminus/viaduct/issues/1
[2] https://github.com/spf13/pflag

+++ directory.go
@@ -30,6 +30,10 @@ func (d Directory) Create() *Directory {
 	d.satisfy()
 
 	log.Println("==> Directory [create]", d.Path)
+	if Config.DryRun {
+		return &d
+	}
+
 	err := os.MkdirAll(d.Path, d.Mode)
 	if err != nil {
 		log.Fatal(err)
@@ -43,6 +47,10 @@ func (d Directory) Delete() *Directory {
 	d.satisfy()
 
 	log.Println("==> Directory [delete]", d.Path)
+	if Config.DryRun {
+		return &d
+	}
+
 	err := os.RemoveAll(d.Path)
 	if err != nil {
 		log.Fatal(err)
